requirements.txt
# General Environment
python-dotenv

# Flask Environment
Flask

#Spotify Environment
Spotipy
